Adjust Destroyable#own() for ES6 promises, refs ibm-js/sdk#5.
Destroyable#destroy() now calls cancel() or destroy() on thenables that are unfulfilled.
However, for this to work, the thenable must have a destroy() or cancel() method, which native
Promises do not.  Perhaps support should just be removed altogether.

Note that previously it called reject(), but for ES6 Promises the reject() method is private.