Make Stateful call custom setters/getters for function properties too.
Also changed the naming of the shadow properties to avoid a conflict where
the getTextDir() property got a shadow property called _getTextDirAttr(),
which looked like a custom getter for the textdir property.
This shouldn't affect existing code assuming it uses this.get("prop")
rather than using the name of the shadow property directly.

Fixes #52 and refs ibm-js/delite#314.